More warning cleanups

+++ src/SQLlib/guid.c
@@ -207,7 +207,7 @@ uint8_t* GetGUIDBinaryEx( CTEXTSTR guid, LOGICAL little_endian )
 	}
 	for( b = n = 0; guid[n]; n++ )
 	{
-		if( char_lookup[guid[n]] < 0 )
+		if( char_lookup[(int)guid[n]] < 0 )
 			continue;
       if( !( b & 1 ) )
 			buf[b / 2] = char_lookup[guid[n]] << 4;
